% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure.fishtank}
\alias{configure.fishtank}
\title{configure fishtank}
\usage{
\method{configure}{fishtank}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
configure fishtank
}
\examples{
configure(item_make(type="fishtank"))
}
