% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure.item}
\alias{configure.item}
\title{Configure item}
\usage{
\method{configure}{item}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
configured item
}
\description{
Configure item
}
\examples{
configure(item_make(type="fishtank"))
}
