% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure.lamp}
\alias{configure.lamp}
\alias{confiugre}
\title{configure lamp}
\usage{
\method{configure}{lamp}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
configured lamp
}
\description{
configure lamp
}
\examples{
configure(item_make(type="lamp"))
}
