% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure}
\alias{configure}
\title{Configure something}
\usage{
configure(x)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
something configured
}
\description{
Configure something
}
\examples{
configure(item_make(type="fishtank"))
}
