% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{item_make}
\alias{item_make}
\title{make a new item}
\usage{
item_make(type = NULL, configure_now = TRUE, ...)
}
\arguments{
\item{type}{}

\item{...}{}
}
\value{
the new item of class input type + item
}
\description{
make a new item
}
\examples{
item_make(type="fishtank")
}
