% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property_make}
\alias{property_make}
\title{make a property}
\usage{
property_make(value = NA, value_class = NULL, name = NULL, ...)
}
\arguments{
\item{value}{}

\item{value_class}{}

\item{name}{}

\item{...}{}
}
\value{
a property
}
\description{
make a property
}
\examples{
property_make()
}
