% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{set2<-.item}
\alias{set2<-.item}
\title{Configure item}
\usage{
\method{set2}{item}(x, property, value_class = NULL) <- value
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
configured item
}
\description{
Configure item
}
\examples{
configure(item_make(type="fishtank"))
}
