% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{set2<-}
\alias{set2<-}
\title{Configure something}
\usage{
set2(x) <- value
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
something configured
}
\description{
Configure something
}
\examples{
configure(item_make(type="fishtank"))
}
